# Add fast pathway for `copy`, `collect`, `tcollect`, and `tcopy` for size-stable operations