--- a/test/test_aqua.jl
+++ b/test/test_aqua.jl
@@ -8,7 +8,10 @@ using Transducers
 Aqua.test_all(
     Transducers;
     # https://github.com/JuliaCollections/DataStructures.jl/pull/511
-    ambiguities = (; exclude = [Base.get, Setfield.set, Setfield.modify, map!]),
+    ambiguities = (; exclude = [Base.get, Setfield.set, Setfield.modify, map!,
+                                # These are needed because of a potential ambiguity with ChainRulesCore's mapfoldl on a Thunk.
+                                # mapreduce is deprecated, so if we remove it, then that would solve this
+                                mapreduce, Core.kwcall]),
     unbound_args = false,  # TODO: make it work
 )