--- a/test/test_aqua.jl
+++ b/test/test_aqua.jl
@@ -14,7 +14,7 @@ push!(to_exclude, mapreduce)
 if isdefined(Core, :kwcall)
     push!(to_exclude, Core.kwcall)
 else
-    push!(to_exclude, geproperty(Base, Symbol("#mapreduce##kw")))
+    push!(to_exclude, getproperty(Base, Symbol("#mapreduce##kw")))
 end
 
 Aqua.test_all(