--- a/test/test_copy.jl
+++ b/test/test_copy.jl
@@ -8,6 +8,10 @@ using Transducers
     @test copy(Map(x -> x^2), [1:3;]) == (1:3) .^ 2
     @test copy(Map(x -> x^2), Set, 1:3) == Set((1:3) .^ 2)
     @test copy(Map(x -> x => x^2), Dict, 2:2) == Dict(2 => 4)
+    @test copy(Map(x -> x-1), Vector, (1, 2, 3)) == [0, 1, 2]
+
+    @test copy(Map(identity), Vector, [1 2; 3 4]) == [1, 3, 2, 4]
+    @test copy(Map(identity), Matrix, [1, 2, 3, 4]) == reshape([1, 2, 3, 4], 4, 1)
 end
 
 @testset "Set" begin